+++ Particle_Transport/Material/material.h
@@ -8,18 +8,18 @@
 #include <string>
 #include <array>
 
-const double AVOGADRO = 6.022e23;
+const float AVOGADRO = 6.022e23;
 
 // Define material (non-radioactive)
 class Material {
     private:
         std::string name;
         short int atomicMass;
-        double density;
-        double atomicDensity;  // atoms/cm^3
+        float density;
+        float atomicDensity;  // atoms/cm^3
         XSTable xstable;
     public:
-        Material(std::string _name, short int _atomicMass, double _density,
+        Material(std::string _name, short int _atomicMass, float _density,
           const XSTable& _xstable) : xstable(_xstable) {
             name = _name;
             atomicMass = _atomicMass;
@@ -29,10 +29,10 @@ class Material {
 
         std::string getName();
         short int getAMass();
-        double getDensity();
-        double getADensity();
-        double getXS(EventType et, const Particle& incP);
-        std::array<double, 3> getEventXSs(const Particle& incP);
+        float getDensity();
+        float getADensity();
+        float getXS(EventType et, const Particle& incP);
+        std::array<float, 3> getEventXSs(const Particle& incP);
         std::array<XSRecord, 3> getEventRecords(const Particle& incP);
 };
