+++ Particle_Transport/Vector3/vector3.h
@@ -1,4 +1,4 @@
-// CURRENTLY NOT USED (Oct. 6, 2025)
+// vector3.h
 
 #ifndef VECTOR3_H
 #define VECTOR3_H
@@ -6,26 +6,26 @@
 #include <cmath>
 
 struct Vector3 {
-    double x, y, z;
+    float x, y, z;
 
-    Vector3(double _x=0.0, double _y=0.0, double _z=0.0) : x(_x), y(_y), z(_z) {}
+    Vector3(float _x=0.0, float _y=0.0, float _z=0.0) : x(_x), y(_y), z(_z) {}
 
     Vector3 operator+(const Vector3& v) const;
     Vector3 operator-(const Vector3& v) const;
-    Vector3 operator*(double a) const;  // Multiply every component by a
-    double dot(const Vector3& v) const;
+    Vector3 operator*(float a) const;  // Multiply every component by a
+    float dot(const Vector3& v) const;
     Vector3 cross(const Vector3& v) const;
     Vector3 operator/(const Vector3& v) const;
     void normalize();
     // Creates randomly chosen orthogonal vector (orthogonal to this Vector3)
     // Uses random u in range [0; 1) for x, y and z
-    Vector3 randomOrthogonalV3(double u_x, double u_y, double u_z) const;
+    Vector3 randomOrthogonalV3(float u_x, float u_y, float u_z) const;
 };
 
 // Is it needed?
 // struct Point3 {
-//     double x, y, z;
-//     Point3(double _x, double _y, double _z) : x(_x), y(_y), z(_z) {}
+//     float x, y, z;
+//     Point3(float _x, float _y, float _z) : x(_x), y(_y), z(_z) {}
 // }
 
 #endif
